pyproject.toml
[tool.poetry]
name = "qdrant-cli"
version = "0.1.0"
description = ""
authors = ["Aaron Collins <aaron@yurts.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyinvoke = "^1.0.4"
qdrant-client = "^1.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
qdrant = 'qdrant_cli:program.run'